Reformatting controller using loadVbox logic
